Updated event lists in View Member and Profile screens
